+++ source/blood/src/gameutil.h
@@ -35,10 +35,10 @@ struct HITINFO {
 
 extern POINT2D baseWall[kMaxWalls];
 extern POINT3D baseSprite[kMaxSprites];
-extern long baseFloor[kMaxSectors];
-extern long baseCeil[kMaxSectors];
-extern long velFloor[kMaxSectors];
-extern long velCeil[kMaxSectors];
+extern int baseFloor[kMaxSectors];
+extern int baseCeil[kMaxSectors];
+extern int velFloor[kMaxSectors];
+extern int velCeil[kMaxSectors];
 extern short gUpperLink[kMaxSectors];
 extern short gLowerLink[kMaxSectors];
 extern HITINFO gHitInfo;
@@ -52,12 +52,12 @@ bool CheckProximityPoint(int nX1, int nY1, int nZ1, int nX2, int nY2, int nZ2, i
 bool CheckProximityWall(int nWall, int x, int y, int nDist);
 int GetWallAngle(int nWall);
 void GetWallNormal(int nWall, int *pX, int *pY);
-bool IntersectRay(long wx, long wy, long wdx, long wdy, long x1, long y1, long z1, long x2, long y2, long z2, long *ix, long *iy, long *iz);
-int HitScan(spritetype *pSprite, int z, int dx, int dy, int dz, unsigned long nMask, int a8);
+bool IntersectRay(int wx, int wy, int wdx, int wdy, int x1, int y1, int z1, int x2, int y2, int z2, int *ix, int *iy, int *iz);
+int HitScan(spritetype *pSprite, int z, int dx, int dy, int dz, unsigned int nMask, int a8);
 int VectorScan(spritetype *pSprite, int nOffset, int nZOffset, int dx, int dy, int dz, int nRange, int ac);
-void GetZRange(spritetype *pSprite, long *ceilZ, long *ceilHit, long *floorZ, long *floorHit, int nDist, unsigned long nMask);
-void GetZRangeAtXYZ(long x, long y, long z, int nSector, long *ceilZ, long *ceilHit, long *floorZ, long *floorHit, int nDist, unsigned long nMask);
+void GetZRange(spritetype *pSprite, int *ceilZ, int *ceilHit, int *floorZ, int *floorHit, int nDist, unsigned int nMask);
+void GetZRangeAtXYZ(int x, int y, int z, int nSector, int *ceilZ, int *ceilHit, int *floorZ, int *floorHit, int nDist, unsigned int nMask);
 int GetDistToLine(int x1, int y1, int x2, int y2, int x3, int y3);
-unsigned int ClipMove(long *x, long *y, long *z, int *nSector, long xv, long yv, int wd, int cd, int fd, unsigned long nMask);
+unsigned int ClipMove(int *x, int *y, int *z, int *nSector, int xv, int yv, int wd, int cd, int fd, unsigned int nMask);
 int GetClosestSectors(int nSector, int x, int y, int nDist, short *pSectors, char *pSectBit);
 int GetClosestSpriteSectors(int nSector, int x, int y, int nDist, short *pSectors, char *pSectBit, short *a8);
